Refactor: extract `StartedMongoSourceTask` from `MongoSourceTask`
KAFKA-308
