+++ src/integrationTest/java/com/mongodb/kafka/connect/source/MongoSourceTaskIntegrationJunkTest.java
@@ -22,8 +22,6 @@ import static com.mongodb.kafka.connect.source.MongoSourceConfig.CONNECTION_URI_
 import static com.mongodb.kafka.connect.source.MongoSourceConfig.DATABASE_CONFIG;
 import static com.mongodb.kafka.connect.source.MongoSourceConfig.FULL_DOCUMENT_CONFIG;
 import static com.mongodb.kafka.connect.source.MongoSourceConfig.PIPELINE_CONFIG;
-import static com.mongodb.kafka.connect.source.SourceTestHelper.TEST_COLLECTION;
-import static com.mongodb.kafka.connect.source.SourceTestHelper.TEST_DATABASE;
 import static com.mongodb.kafka.connect.util.jmx.internal.MBeanServerUtils.getMBeanAttributes;
 import static java.lang.String.format;
 import static java.util.Collections.singletonMap;
@@ -33,6 +31,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -64,8 +63,15 @@ import com.mongodb.client.model.changestream.FullDocument;
 import com.mongodb.event.CommandFailedEvent;
 import com.mongodb.event.CommandSucceededEvent;
 
+import com.mongodb.kafka.connect.util.jmx.SourceTaskStatistics;
+
+/**
+ * This class contains tests that are supposed to be unit tests, but because of how these tests were
+ * written originally, they became integration tests as a result of refactoring {@link
+ * MongoSourceConfig}, and need to be completely rewritten to remain unit tests.
+ */
 @ExtendWith(MockitoExtension.class)
-class MongoSourceTaskTest {
+class MongoSourceTaskIntegrationJunkTest {
 
   @Mock private MongoClient mongoClient;
   @Mock private MongoDatabase mongoDatabase;
@@ -76,6 +82,8 @@ class MongoSourceTaskTest {
   @Mock private SourceTaskContext context;
   @Mock private OffsetStorageReader offsetStorageReader;
 
+  private static final String TEST_DATABASE = "myDB";
+  private static final String TEST_COLLECTION = "myColl";
   private static final BsonDocument RESUME_TOKEN = BsonDocument.parse("{resume: 'token'}");
   private static final Map<String, Object> OFFSET = singletonMap("_id", RESUME_TOKEN.toJson());
 
@@ -88,6 +96,7 @@ class MongoSourceTaskTest {
     cfgMap.put(DATABASE_CONFIG, TEST_DATABASE);
     cfgMap.put(COLLECTION_CONFIG, TEST_COLLECTION);
     MongoSourceConfig cfg = new MongoSourceConfig(cfgMap);
+    task.start(cfgMap);
 
     when(mongoClient.getDatabase(TEST_DATABASE)).thenReturn(mongoDatabase);
     when(mongoDatabase.getCollection(TEST_COLLECTION)).thenReturn(mongoCollection);
@@ -95,7 +104,7 @@ class MongoSourceTaskTest {
     when(changeStreamIterable.withDocumentClass(RawBsonDocument.class)).thenReturn(mongoIterable);
     when(mongoIterable.cursor()).thenReturn(mongoCursor);
 
-    task.createCursor(cfg, mongoClient);
+    task.startedTask().createCursor(cfg, mongoClient);
 
     verify(mongoClient, times(1)).getDatabase(TEST_DATABASE);
     verify(mongoDatabase, times(1)).getCollection(TEST_COLLECTION);
@@ -114,7 +123,7 @@ class MongoSourceTaskTest {
     when(changeStreamIterable.withDocumentClass(RawBsonDocument.class)).thenReturn(mongoIterable);
     when(mongoIterable.cursor()).thenReturn(mongoCursor);
 
-    task.createCursor(cfg, mongoClient);
+    task.startedTask().createCursor(cfg, mongoClient);
 
     verify(mongoClient, times(1)).getDatabase(TEST_DATABASE);
     verify(mongoDatabase, times(1)).getCollection(TEST_COLLECTION);
@@ -146,7 +155,7 @@ class MongoSourceTaskTest {
 
     task.initialize(context);
     when(context.offsetStorageReader()).thenReturn(offsetStorageReader);
-    when(offsetStorageReader.offset(task.createPartitionMap(cfg))).thenReturn(OFFSET);
+    when(offsetStorageReader.offset(MongoSourceTask.createPartitionMap(cfg))).thenReturn(OFFSET);
 
     when(mongoClient.getDatabase(TEST_DATABASE)).thenReturn(mongoDatabase);
     when(mongoDatabase.getCollection(TEST_COLLECTION)).thenReturn(mongoCollection);
@@ -158,7 +167,7 @@ class MongoSourceTaskTest {
     when(changeStreamIterable.withDocumentClass(RawBsonDocument.class)).thenReturn(mongoIterable);
     when(mongoIterable.cursor()).thenReturn(mongoCursor);
 
-    task.createCursor(cfg, mongoClient);
+    task.startedTask().createCursor(cfg, mongoClient);
 
     verify(mongoClient, times(1)).getDatabase(TEST_DATABASE);
     verify(mongoDatabase, times(1)).getCollection(TEST_COLLECTION);
@@ -169,6 +178,8 @@ class MongoSourceTaskTest {
     verify(changeStreamIterable, times(1)).startAfter(RESUME_TOKEN);
     verify(changeStreamIterable, times(1)).withDocumentClass(RawBsonDocument.class);
     verify(mongoIterable, times(1)).cursor();
+
+    task.stop();
   }
 
   @Test
@@ -179,13 +190,14 @@ class MongoSourceTaskTest {
     cfgMap.put(CONNECTION_URI_CONFIG, "mongodb://localhost");
     cfgMap.put(DATABASE_CONFIG, TEST_DATABASE);
     MongoSourceConfig cfg = new MongoSourceConfig(cfgMap);
+    task.start(cfgMap);
 
     when(mongoClient.getDatabase(TEST_DATABASE)).thenReturn(mongoDatabase);
     when(mongoDatabase.watch()).thenReturn(changeStreamIterable);
     when(changeStreamIterable.withDocumentClass(RawBsonDocument.class)).thenReturn(mongoIterable);
     when(mongoIterable.cursor()).thenReturn(mongoCursor);
 
-    task.createCursor(cfg, mongoClient);
+    task.startedTask().createCursor(cfg, mongoClient);
 
     verify(mongoClient, times(1)).getDatabase(TEST_DATABASE);
     verify(mongoDatabase, times(1)).watch();
@@ -202,7 +214,7 @@ class MongoSourceTaskTest {
     when(changeStreamIterable.withDocumentClass(RawBsonDocument.class)).thenReturn(mongoIterable);
     when(mongoIterable.cursor()).thenReturn(mongoCursor);
 
-    task.createCursor(cfg, mongoClient);
+    task.startedTask().createCursor(cfg, mongoClient);
 
     verify(mongoClient, times(1)).getDatabase(TEST_DATABASE);
     verify(mongoDatabase, times(1)).watch(cfg.getPipeline().get());
@@ -233,7 +245,7 @@ class MongoSourceTaskTest {
 
     task.initialize(context);
     when(context.offsetStorageReader()).thenReturn(offsetStorageReader);
-    when(offsetStorageReader.offset(task.createPartitionMap(cfg))).thenReturn(OFFSET);
+    when(offsetStorageReader.offset(MongoSourceTask.createPartitionMap(cfg))).thenReturn(OFFSET);
 
     when(mongoClient.getDatabase(TEST_DATABASE)).thenReturn(mongoDatabase);
     when(mongoDatabase.watch(cfg.getPipeline().get())).thenReturn(changeStreamIterable);
@@ -244,7 +256,7 @@ class MongoSourceTaskTest {
     when(changeStreamIterable.withDocumentClass(RawBsonDocument.class)).thenReturn(mongoIterable);
     when(mongoIterable.cursor()).thenReturn(mongoCursor);
 
-    task.createCursor(cfg, mongoClient);
+    task.startedTask().createCursor(cfg, mongoClient);
 
     verify(mongoClient, times(1)).getDatabase(TEST_DATABASE);
     verify(mongoDatabase, times(1)).watch(cfg.getPipeline().get());
@@ -254,6 +266,8 @@ class MongoSourceTaskTest {
     verify(changeStreamIterable, times(1)).startAfter(RESUME_TOKEN);
     verify(changeStreamIterable, times(1)).withDocumentClass(RawBsonDocument.class);
     verify(mongoIterable, times(1)).cursor();
+
+    task.stop();
   }
 
   @Test
@@ -263,12 +277,13 @@ class MongoSourceTaskTest {
     Map<String, String> cfgMap = new HashMap<>();
     cfgMap.put(CONNECTION_URI_CONFIG, "mongodb://localhost");
     MongoSourceConfig cfg = new MongoSourceConfig(cfgMap);
+    task.start(cfgMap);
 
     when(mongoClient.watch()).thenReturn(changeStreamIterable);
     when(changeStreamIterable.withDocumentClass(RawBsonDocument.class)).thenReturn(mongoIterable);
     when(mongoIterable.cursor()).thenReturn(mongoCursor);
 
-    task.createCursor(cfg, mongoClient);
+    task.startedTask().createCursor(cfg, mongoClient);
 
     verify(mongoClient, times(1)).watch();
     verify(changeStreamIterable, times(1)).withDocumentClass(RawBsonDocument.class);
@@ -283,7 +298,7 @@ class MongoSourceTaskTest {
     when(changeStreamIterable.withDocumentClass(RawBsonDocument.class)).thenReturn(mongoIterable);
     when(mongoIterable.cursor()).thenReturn(mongoCursor);
 
-    task.createCursor(cfg, mongoClient);
+    task.startedTask().createCursor(cfg, mongoClient);
 
     verify(mongoClient, times(1)).watch(cfg.getPipeline().get());
     verify(changeStreamIterable, times(1)).withDocumentClass(RawBsonDocument.class);
@@ -312,7 +327,7 @@ class MongoSourceTaskTest {
 
     task.initialize(context);
     when(context.offsetStorageReader()).thenReturn(offsetStorageReader);
-    when(offsetStorageReader.offset(task.createPartitionMap(cfg))).thenReturn(OFFSET);
+    when(offsetStorageReader.offset(MongoSourceTask.createPartitionMap(cfg))).thenReturn(OFFSET);
 
     when(mongoClient.watch(cfg.getPipeline().get())).thenReturn(changeStreamIterable);
     when(changeStreamIterable.batchSize(101)).thenReturn(changeStreamIterable);
@@ -322,7 +337,7 @@ class MongoSourceTaskTest {
     when(changeStreamIterable.withDocumentClass(RawBsonDocument.class)).thenReturn(mongoIterable);
     when(mongoIterable.cursor()).thenReturn(mongoCursor);
 
-    task.createCursor(cfg, mongoClient);
+    task.startedTask().createCursor(cfg, mongoClient);
 
     verify(mongoClient, times(1)).watch(cfg.getPipeline().get());
     verify(changeStreamIterable, times(1)).batchSize(101);
@@ -331,30 +346,30 @@ class MongoSourceTaskTest {
     verify(changeStreamIterable, times(1)).startAfter(RESUME_TOKEN);
     verify(changeStreamIterable, times(1)).withDocumentClass(RawBsonDocument.class);
     verify(mongoIterable, times(1)).cursor();
+
+    task.stop();
   }
 
   @Test
   @DisplayName("test handles legacy offsets")
   void testHandlesLegacyOffsets() {
-    MongoSourceTask task = new MongoSourceTask();
     Map<String, String> cfgMap = new HashMap<>();
     cfgMap.put(CONNECTION_URI_CONFIG, "mongodb://localhost");
     cfgMap.put(DATABASE_CONFIG, TEST_DATABASE);
     cfgMap.put(COLLECTION_CONFIG, TEST_COLLECTION);
     MongoSourceConfig cfg = new MongoSourceConfig(cfgMap);
 
-    task.initialize(context);
     when(context.offsetStorageReader()).thenReturn(offsetStorageReader);
-    when(offsetStorageReader.offset(task.createPartitionMap(cfg))).thenReturn(null);
-    when(offsetStorageReader.offset(task.createLegacyPartitionMap(cfg))).thenReturn(OFFSET);
+    when(offsetStorageReader.offset(MongoSourceTask.createPartitionMap(cfg))).thenReturn(null);
+    when(offsetStorageReader.offset(MongoSourceTask.createLegacyPartitionMap(cfg)))
+        .thenReturn(OFFSET);
 
-    assertEquals(OFFSET, task.getOffset(cfg));
+    assertEquals(OFFSET, MongoSourceTask.getOffset(context, cfg));
   }
 
   @Test
   @DisplayName("test creates the expected partition map")
   void testCreatesTheExpectedPartitionMap() {
-    MongoSourceTask task = new MongoSourceTask();
     Map<String, String> cfgMap = new HashMap<>();
     cfgMap.put(CONNECTION_URI_CONFIG, "mongodb+srv://user:password@localhost/");
     cfgMap.put(DATABASE_CONFIG, TEST_DATABASE);
@@ -363,31 +378,33 @@ class MongoSourceTaskTest {
 
     assertEquals(
         format("mongodb+srv://localhost/%s.%s", TEST_DATABASE, TEST_COLLECTION),
-        task.createDefaultPartitionName(cfg));
+        MongoSourceTask.createDefaultPartitionName(cfg));
     assertEquals(
         format("mongodb+srv://user:password@localhost//%s.%s", TEST_DATABASE, TEST_COLLECTION),
-        task.createLegacyPartitionName(cfg));
+        MongoSourceTask.createLegacyPartitionName(cfg));
 
     cfgMap.put(CONNECTION_URI_CONFIG, "mongodb://localhost/");
     cfg = new MongoSourceConfig(cfgMap);
     assertEquals(
         format("mongodb://localhost/%s.%s", TEST_DATABASE, TEST_COLLECTION),
-        task.createDefaultPartitionName(cfg));
+        MongoSourceTask.createDefaultPartitionName(cfg));
     assertEquals(
         format("mongodb://localhost//%s.%s", TEST_DATABASE, TEST_COLLECTION),
-        task.createLegacyPartitionName(cfg));
+        MongoSourceTask.createLegacyPartitionName(cfg));
 
     cfgMap.remove(COLLECTION_CONFIG);
     cfg = new MongoSourceConfig(cfgMap);
     assertEquals(
-        format("mongodb://localhost/%s", TEST_DATABASE), task.createDefaultPartitionName(cfg));
+        format("mongodb://localhost/%s", TEST_DATABASE),
+        MongoSourceTask.createDefaultPartitionName(cfg));
     assertEquals(
-        format("mongodb://localhost//%s.", TEST_DATABASE), task.createLegacyPartitionName(cfg));
+        format("mongodb://localhost//%s.", TEST_DATABASE),
+        MongoSourceTask.createLegacyPartitionName(cfg));
 
     cfgMap.remove(DATABASE_CONFIG);
     cfg = new MongoSourceConfig(cfgMap);
-    assertEquals("mongodb://localhost/", task.createDefaultPartitionName(cfg));
-    assertEquals("mongodb://localhost//.", task.createLegacyPartitionName(cfg));
+    assertEquals("mongodb://localhost/", MongoSourceTask.createDefaultPartitionName(cfg));
+    assertEquals("mongodb://localhost//.", MongoSourceTask.createLegacyPartitionName(cfg));
   }
 
   @Test
@@ -396,7 +413,7 @@ class MongoSourceTaskTest {
     String mBeanName =
         "com.mongodb.kafka.connect:type=source-task-metrics,task=source-task-change-stream-unknown";
     MongoSourceTask task = new MongoSourceTask();
-    task.initializeStatistics(false);
+    task.start(Collections.emptyMap());
 
     task.commitRecord(null, new RecordMetadata(null, 0, 0, 0, 0L, 0, 0));
 
@@ -421,10 +438,10 @@ class MongoSourceTaskTest {
     String mBeanName =
         "com.mongodb.kafka.connect:type=source-task-metrics,task=source-task-change-stream-unknown";
 
-    MongoSourceTask getmoreSuccessTask = new MongoSourceTask();
-    getmoreSuccessTask.initializeStatistics(false);
-    getmoreSuccessTask.mongoCommandSucceeded(
-        new CommandSucceededEvent(0, null, "getMore", new BsonDocument(), 100000000));
+    SourceTaskStatistics stats = new SourceTaskStatistics(mBeanName);
+    stats.register();
+    MongoSourceTask.mongoCommandSucceeded(
+        new CommandSucceededEvent(0, null, "getMore", new BsonDocument(), 100000000), stats);
     for (Map<String, Long> attrs : getMBeanAttributes(mBeanName).values()) {
       assertEquals(1, attrs.get("getmore-commands-successful"));
       assertEquals(100, attrs.get("getmore-commands-successful-duration-ms"));
@@ -432,12 +449,12 @@ class MongoSourceTaskTest {
       assertEquals(1, attrs.get("getmore-commands-successful-duration-over-10-ms"));
       assertEquals(4, attrs.values().stream().filter(v -> v != 0).count());
     }
-    getmoreSuccessTask.stop();
+    stats.unregister();
 
-    MongoSourceTask initiatingSuccessTask = new MongoSourceTask();
-    initiatingSuccessTask.initializeStatistics(false);
-    initiatingSuccessTask.mongoCommandSucceeded(
-        new CommandSucceededEvent(0, null, "aggregate", new BsonDocument(), 100000000));
+    stats = new SourceTaskStatistics(mBeanName);
+    stats.register();
+    MongoSourceTask.mongoCommandSucceeded(
+        new CommandSucceededEvent(0, null, "aggregate", new BsonDocument(), 100000000), stats);
     for (Map<String, Long> attrs : getMBeanAttributes(mBeanName).values()) {
       assertEquals(1, attrs.get("initial-commands-successful"));
       assertEquals(100, attrs.get("initial-commands-successful-duration-ms"));
@@ -445,12 +462,12 @@ class MongoSourceTaskTest {
       assertEquals(1, attrs.get("initial-commands-successful-duration-over-10-ms"));
       assertEquals(4, attrs.values().stream().filter(v -> v != 0).count());
     }
-    initiatingSuccessTask.stop();
+    stats.unregister();
 
-    MongoSourceTask getmoreFailedTask = new MongoSourceTask();
-    getmoreFailedTask.initializeStatistics(false);
-    getmoreFailedTask.mongoCommandFailed(
-        new CommandFailedEvent(0, null, "getMore", 100000000, null));
+    stats = new SourceTaskStatistics(mBeanName);
+    stats.register();
+    MongoSourceTask.mongoCommandFailed(
+        new CommandFailedEvent(0, null, "getMore", 100000000, null), stats);
     for (Map<String, Long> attrs : getMBeanAttributes(mBeanName).values()) {
       assertEquals(1, attrs.get("getmore-commands-failed"));
       assertEquals(100, attrs.get("getmore-commands-failed-duration-ms"));
@@ -458,12 +475,12 @@ class MongoSourceTaskTest {
       assertEquals(1, attrs.get("getmore-commands-failed-duration-over-10-ms"));
       assertEquals(4, attrs.values().stream().filter(v -> v != 0).count());
     }
-    getmoreFailedTask.stop();
+    stats.unregister();
 
-    MongoSourceTask initiatingFailedTask = new MongoSourceTask();
-    initiatingFailedTask.initializeStatistics(false);
-    initiatingFailedTask.mongoCommandFailed(
-        new CommandFailedEvent(0, null, "aggregate", 100000000, null));
+    stats = new SourceTaskStatistics(mBeanName);
+    stats.register();
+    MongoSourceTask.mongoCommandFailed(
+        new CommandFailedEvent(0, null, "aggregate", 100000000, null), stats);
     for (Map<String, Long> attrs : getMBeanAttributes(mBeanName).values()) {
       assertEquals(1, attrs.get("initial-commands-failed"));
       assertEquals(100, attrs.get("initial-commands-failed-duration-ms"));
@@ -471,7 +488,7 @@ class MongoSourceTaskTest {
       assertEquals(1, attrs.get("initial-commands-failed-duration-over-10-ms"));
       assertEquals(4, attrs.values().stream().filter(v -> v != 0).count());
     }
-    initiatingFailedTask.stop();
+    stats.unregister();
   }
 
   private void resetMocks() {
